services/brain/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.4
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1
python-dotenv==1.0.0
httpx==0.25.2
paho-mqtt==1.6.1
aiosqlite==0.19.0